feat:add sgd optimizer

+++ src/deeplearning/loss/mse_loss.h
@@ -5,10 +5,11 @@ namespace deeplearning {
 class MSELoss : public LossFunction {
 public:
   double Loss(double target, double output) override {
-    return (1 / 2) * (target - output) * (target - output);
+    return (double)(1.0 / 2.0) * (double)(target - output) *
+           (double)(target - output);
   }
   double DerivLoss(double target, double output) override {
-    return -2 * (target - output);
+    return -2.0 * (target - output);
   }
   LossType GetLossType() override { return LOSS_MSE; }
 };
